Changed the ListView choice mode to CHOICE_MODE_MULTI. Changed the background of importer_detail_listview
CHOICE_MODE_MULTI requires a little more work to use an ActionMode with but it allows the onItemClick to be called during the actionmode. This allows me to set the checked state of the checkbox without some ugly hacks. The background was changed because it was redundant indication of a listitem checked state by the UI